+++ apps/faces-example1/src/main/java/org/apache/struts/webapp/example/RegistrationBacking.java
@@ -23,6 +23,8 @@
 package org.apache.struts.webapp.example;
 
 
+import javax.faces.bean.ManagedBean;
+import javax.faces.bean.RequestScoped;
 import javax.faces.component.UIData;
 import javax.faces.context.FacesContext;
 
@@ -37,6 +39,8 @@ import org.slf4j.LoggerFactory;
  * <p>Backing bean for the <code>registration.jsp</code> page.</p>
  */
 
+@ManagedBean
+@RequestScoped
 public class RegistrationBacking extends AbstractBacking {